convert fujinet protocol library to Z80
convert the fujinet protocol library to Z80 assembler.
This allows us to embed the code into a driver

+++ lib/include/fujinet.h
@@ -14,7 +14,7 @@ struct fujinet_dcb {
     uint16_t buffer_bytes; // data length in bytes
     uint8_t *response;
     uint16_t response_bytes; // data length in bytes
-    uint8_t timeout;   // seconds
+    uint16_t timeout;   // 10ths of seconds
 };
 
 #define FUJINET_TIMEOUT 15
